simplify method typing
