+++ packages/insomnia/src/ui/components/dropdowns/grpc-method-dropdown/grpc-method-dropdown.tsx
@@ -1,11 +1,11 @@
-import { MethodDefinition } from '@grpc/grpc-js';
-import React, { Fragment, FunctionComponent, useMemo } from 'react';
+import React, { Fragment, FunctionComponent } from 'react';
 import styled from 'styled-components';
 
 import {
   getGrpcPathSegments,
   getShortGrpcPath,
   groupGrpcMethodsByPackage,
+  GrpcMethodInfo,
   NO_PACKAGE_KEY,
 } from '../../../../common/grpc-paths';
 import { Dropdown } from '../../base/dropdown/dropdown';
@@ -36,8 +36,8 @@ const DropdownMethodButtonLabel = styled.div({
 
 interface Props {
   disabled?: boolean;
-  methods: MethodDefinition<any, any>[];
-  selectedMethod?: MethodDefinition<any, any>;
+  methods: GrpcMethodInfo[];
+  selectedMethod?: GrpcMethodInfo;
   handleChange: (arg0: string) => void;
   handleChangeProtoFile: () => void;
 }
@@ -54,15 +54,8 @@ export const GrpcMethodDropdown: FunctionComponent<Props> = ({
   handleChangeProtoFile,
 }) => {
   const groupedByPkg = groupGrpcMethodsByPackage(methods);
-  const useLabel = (fullPath?: string) =>
-    useMemo(() => {
-      if (fullPath) {
-        const segments = getGrpcPathSegments(fullPath);
-        return getShortGrpcPath(segments, fullPath);
-      }
+  const selectedPath = selectedMethod?.fullPath;
 
-      return 'Select Method';
-    }, [fullPath]);
   return (
     <Dropdown
       className="tall wide"
@@ -70,8 +63,8 @@ export const GrpcMethodDropdown: FunctionComponent<Props> = ({
       <DropdownButton
         buttonClass={DropdownMethodButton}
       >
-        <Tooltip message={selectedMethod?.path || 'Select Method'} position="bottom" delay={500}>
-          {useLabel(selectedMethod?.path)}
+        <Tooltip message={selectedPath || 'Select Method'} position="bottom" delay={500}>
+          {!selectedPath ? 'Select Method' : getShortGrpcPath(getGrpcPathSegments(selectedPath), selectedPath)}
           <i className="fa fa-caret-down pad-left-sm" />
         </Tooltip>
       </DropdownButton>
@@ -94,7 +87,7 @@ export const GrpcMethodDropdown: FunctionComponent<Props> = ({
               key={fullPath}
               onClick={() => handleChange(fullPath)}
               disabled={disabled}
-              selected={fullPath === selectedMethod?.path}
+              selected={fullPath === selectedPath}
             >
               <Tooltip message={fullPath} position="right" delay={500}>
                 <DropdownMethodButtonLabel><GrpcMethodTag methodType={type} /> {getShortGrpcPath(segments, fullPath)}</DropdownMethodButtonLabel>
